No ramo develop
- Função de Exclusão Global;
- Fabricante Seeers;

- Mudanças a serem submetidas:
	modified:   app/Http/Controllers/FabricanteController.php
	new file:   database/factories/FabricanteFactory.php
	modified:   database/seeds/DatabaseSeeder.php
	new file:   database/seeds/FabricanteSeeder.php
	modified:   database/seeds/UserSeeder.php
	modified:   resources/views/fabricantes/index.blade.php
	modified:   resources/views/vendor/adminlte/page.blade.php

- Arquivos não monitorados:
	resources/views/vendor/adminlte/auth/
	resources/views/vendor/adminlte/master.blade.php
	resources/views/vendor/adminlte/partials/
	resources/views/vendor/adminlte/plugins.blade.php
	resources/views/vendor/datatables/

@rodrixcornell

+++ resources/views/fabricantes/index.blade.php
@@ -26,12 +26,36 @@
 
 <script>
     function excluir(rota) {
-        axios.delete(rota)
-        .then((data) => {
-            alert('Sucesso ao apagar')
-        })
-        .catch((err) => {
-            alert('Erro ao apagar')
+        // console.log(window.LaravelDataTables)
+        // console.log(Object.keys(window.LaravelDataTables)[0])
+        Swal.fire({
+            title: 'Atenção?',
+            text: "Deseja mesmo excluir?",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Sim',
+            cancelButtonText: 'Não'
+            }).then((result) => {
+            if (result.isConfirmed) {
+                axios.delete(rota)
+                    .then((data) => {
+                        $('#' + Object.keys(window.LaravelDataTables)[0]).DataTable().ajax.reload()
+                        Swal.fire(
+                            'Perfeito!',
+                            'Excluido com sucesso.',
+                            'success'
+                        )
+                    })
+                    .catch((err) => {
+                        Swal.fire(
+                            'Ops!',
+                            'Error ao excluir.',
+                            'success'
+                        )
+                    })
+            }
         })
     }
 </script>
